refactor: fsd에 따라 도메인 구분에 따른 폴더 분리 진행
