+++ src/features/add-sequence/ui/SequenceAddItem/sequenceAdd.css.ts
