--- src/pages/popup/Sequence/module/sequence/useHandleSequencePage.ts
@@ -1,53 +1,0 @@
-import { SequenceFunnelProps } from "@/pages/popup/pages/SequenceFunnel/SequenceFunnel";
-import useRouter from "@/shared/hooks/useRouter";
-import { useState } from "react";
-import { addSequence } from "../../../../../entities/sequence/model/sequence-store";
-
-type Mode = "VIEW" | "ADD";
-
-type OmitOnNext = Omit<SequenceFunnelProps, "onNext" | "sequenceList">;
-type HandleSequencePageProps = OmitOnNext;
-
-const useHandleSequencePage = ({ swaggerTitle }: HandleSequencePageProps) => {
-  const router = useRouter();
-  const [mode, setMode] = useState<Mode>("VIEW");
-
-  const onClickAdd = () => {
-    setMode("ADD");
-  };
-
-  const onClickClose = () => {
-    setMode("VIEW");
-  };
-
-  // USER INTERACTION
-  const [addTitle, setAddTitle] = useState<string>("");
-
-  const onChangeTitle = (title: string) => {
-    setAddTitle(title);
-  };
-
-  const onPressEnter = (title: string) => {
-    addSequence(swaggerTitle, { title, iconType: "LOADING", ApiList: [] });
-    setMode("VIEW");
-    setAddTitle("");
-  };
-
-  const onClickSequence = (id: number, title: string) => {
-    router.push(`/sequence/${id}`, { swaggerTitle, title });
-  };
-
-  return {
-    modeSate: {
-      mode,
-      onClickAdd,
-      onClickClose,
-    },
-    addTitle,
-    onChangeTitle,
-    onPressEnter,
-    onClickSequence,
-  };
-};
-
-export default useHandleSequencePage;
