--- src/pages/popup/Sequence/store/api.ts
@@ -1,14 +1,0 @@
-import { create } from "zustand";
-import { APIWithOrder } from "../../../../entities/sequence/model/sequence-store";
-
-interface apiStore {
-  apis: APIWithOrder[];
-  setApis: (apis: APIWithOrder[]) => void;
-}
-
-const useAPIStore = create<apiStore>((set) => ({
-  apis: [],
-  setApis: (apis: APIWithOrder[]) => set({ apis }),
-}));
-
-export default useAPIStore;
