+++ src/pages/popup/index.tsx
@@ -4,9 +4,9 @@ import { createRoot } from "react-dom/client";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import refreshOnUpdate from "virtual:reload-on-update-in-view";
 import "../../shared/ui/styles/reset.css.js";
-import MultipleTestPage from "./Sequence/pages/Multiple/MultipleTestPage.js";
 import "./index.css";
 import Popup from "./pages/ApiList/ui/ApiList.js";
+import MultipleTestPage from "./pages/Multiple/MultipleTestPage.js";
 import Request from "./pages/Request/Request";
 import ScenarioFunnel from "./pages/ScenarioFunnel/ScenarioFunnel.js";
 import SequenceFunnel from "./pages/SequenceFunnel/SequenceFunnel.js";
