+++ src/pages/popup/pages/Scenario/scenario.css.ts
