+++ src/pages/popup/pages/Sequence/sequence.css.ts
