+++ src/widgets/scenario-list/ui/list/ScenarioList.tsx
@@ -1,0 +1,20 @@
+import { APIWithOrder } from "@/entities/sequence/model/sequence-store";
+import ScenarioItem from "@/widgets/scenario-list/ui/item/ScenarioItem";
+
+interface ScenarioListProps {
+  scenarioList: APIWithOrder[];
+  onClickAPI: (idx: number) => void;
+}
+
+export const ScenarioList = ({
+  scenarioList,
+  onClickAPI,
+}: ScenarioListProps) => {
+  return (
+    <>
+      {scenarioList.map((api, idx) => (
+        <ScenarioItem key={api.key} api={api} onClick={() => onClickAPI(idx)} />
+      ))}
+    </>
+  );
+};
