+++ src/widgets/sequence-list/ui/item/sequenceItem.css.ts
